+++ nrv/utils/_saving_handler.py
@@ -14,14 +14,15 @@ from ..nmod._myelinated import *
 from ..nmod._unmyelinated import *
 from ..backend._log_interface import rise_warning
 
-###############################################################
-#########################  Loaders  ###########################
-###############################################################
+# --------- #
+#  Loaders  #
+# --------- #
+
 
 
 def load_any_fascicle(
-    data, extracel_context=False, intracel_context=False, rec_context=False
-):
+    data:str|dict, extracel_context:bool=False, intracel_context:bool=False, rec_context:bool=False
+)->fascicle:
     """
     generate any kind of fascicle from a dictionary or a json file
 
@@ -54,7 +55,7 @@ def load_any_fascicle(
 
 def load_any_axon(
     data, extracel_context=False, intracel_context=False, rec_context=False
-):
+)->axon:
     """
     generate any kind of axon from a dictionary or a json file
 
@@ -88,3 +89,4 @@ def load_any_axon(
         return ax, ax.recorder
     else:
         return ax
+