+++ nrv/utils/geom/_ellipse.py
@@ -11,7 +11,7 @@ class Ellipse(CShape):
     Represents an ellipse with a center, semi-major axis, and semi-minor axis.
     """
 
-    def __init__(self, center: tuple[float, float], radius:tuple[float, float], rot:float=None, degree:bool=False):
+    def __init__(self, center: tuple[float, float]=(0,0), radius:tuple[float, float]=(10,10), rot:float=None, degree:bool=False):
         """
         Initialize the Ellipse with a center and axes.
 
@@ -19,14 +19,11 @@ class Ellipse(CShape):
         :param semi_major_axis: The length of the semi-major axis.
         :param r2: The length of the semi-minor axis.
         """
-        super().__init__(center, radius)
+        super().__init__(center, radius, rot=rot)
         assert np.iterable(radius), "Ellipse radius must be iterable (of lenght at least equal to 2)"
-        self.rot = 0
         self.r1 = self.radius[0]
         self.r2 = self.radius[1]
 
-        if rot is not None:
-            self.rotate(rot, degree)
 
 
     @property
@@ -37,20 +34,17 @@ class Ellipse(CShape):
         return np.array(self.radius, dtype=float)
 
     @property
-    def x(self)->float:
-        return 0
-    
-    @property
-    def y(self)->float:
-        return self.center[0]
-
-    @property
-    def z(self)->float:
-        return self.center[1]
-    
-    @property
     def area(self)->float:
         return np.pi * self.r1 * self.r2
+    
+    @property
+    def bbox_size(self)->tuple[float, float]:
+        return (
+            np.hypot(2 * self.r1 * np.cos(-self.rot),
+            2 * self.r2 * np.sin(-self.rot)),
+            np.hypot(2 * self.r1 * np.sin(-self.rot),
+            2 * self.r2 * np.cos(-self.rot)),
+        )
 
     @property
     def rot_mat(self)->np.ndarray:
@@ -105,7 +99,7 @@ class Ellipse(CShape):
     def rotate(self, angle:float, degree:bool=False):
         if degree:
             angle = to_nrv_unit(angle, "deg")
-        self.rot += angle
+        self.rot = (self.rot + angle)%(2*np.pi)
 
 
     def get_trace(self, n_theta=100) -> tuple[list[float], list[float]]: