+++ tests/unitary_tests/059_Fascicle_extra_stim.py
@@ -28,7 +28,7 @@ if __name__ == '__main__':
 
     # Fascicle declaration
     fascicle_1 = nrv.fascicle()
-    fascicle_1.load_fascicle_configuration(source_file)
+    fascicle_1.load(source_file)
     fascicle_1.define_length(L)
     fascicle_1.set_ID(59)
     # extra cellular stimulation